requirements.txt
# Core requirements
streamlit==1.50.0
opencv-python==4.12.0
numpy>=2.2.6
pillow>=11.3.0
ultralytics>=8.0.0
pandas>=2.3.3

# Database
sqlite3

# UI and visualization
plotly>=5.0.0

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0

# Utilities
python-dotenv>=0.19.0
requests>=2.32.5
pyyaml>=6.0.0